requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
asyncio==3.4.3
numpy==1.16.3
starlette==0.12.0
uvicorn==0.7.1
python-multipart==0.0.5
deepstack-python==0.4
pillow
